check tests refactor
